scripts/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.2.0=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39hb18efdd_2
asttokens=2.0.8=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bayesian-optimization=1.2.0=pypi_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
boto3=1.24.75=pyhd8ed1ab_0
botocore=1.27.75=pyhd8ed1ab_0
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py39hb18efdd_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.10.11=hca03da5_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pypi_0
category_encoders=2.5.0=py39h86d0a89_1
certifi=2022.9.24=py39hca03da5_0
cffi=1.15.1=py39h04d3946_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=py39h2804cbe_0
cloudpickle=2.2.0=pypi_0
colorama=0.4.5=pyhd8ed1ab_0
contourpy=1.0.5=py39haaf3ac1_0
cryptography=37.0.4=py39h4cb6239_0
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.3=py39h3c22d25_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
et_xmlfile=1.1.0=py39hca03da5_0
executing=1.0.0=pyhd8ed1ab_0
filelock=3.8.0=pypi_0
flask=2.2.2=pyhd8ed1ab_0
flatbuffers=2.0.7=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.37.2=py39h02fc5c5_0
freetype=2.12.1=hd633e50_0
gast=0.4.0=pypi_0
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=py39h518bade_0
gym=0.26.0=pypi_0
gym-notices=0.0.8=pypi_0
h5py=3.6.0=nompi_py39hd982b79_100
hdf5=1.12.1=nompi_hf9525e8_104
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
imbalanced-learn=0.9.1=pyhd8ed1ab_1
importlib-metadata=4.11.4=py39h2804cbe_0
importlib_resources=5.9.0=pyhd8ed1ab_0
ipykernel=6.15.3=pyh736e0ef_0
ipython=8.5.0=pyhd1c38e8_1
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.2=pyhd8ed1ab_1
itsdangerous=2.1.2=pyhd8ed1ab_0
jedi=0.18.1=pyhd8ed1ab_2
jinja2=3.1.2=pyhd8ed1ab_1
jmespath=1.0.1=pyhd8ed1ab_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter=1.0.0=py39h2804cbe_7
jupyter_client=7.3.5=pyhd8ed1ab_0
jupyter_console=6.4.4=pyhd8ed1ab_0
jupyter_core=4.11.1=py39h2804cbe_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=3.0.3=pyhd8ed1ab_0
kaggle=1.5.12=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=py39hab5e169_0
krb5=1.19.3=hf9b2bbe_0
lcms2=2.12=had6a04f_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=16_osxarm64_openblas
libclang=14.0.6=pypi_0
libcurl=7.83.1=h2fcd78c_0
libcxx=14.0.6=h2692d47_0
libdeflate=1.14=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_25
libgfortran5=11.3.0=hdaf2cc0_25
libiconv=1.16=h642e427_0
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.47.0=h232270b_1
libopenblas=0.3.21=openmp_hc731615_3
libpng=1.6.38=h76d750c_0
libprotobuf=3.19.4=hccf11d3_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.39.3=h76d750c_0
libssh2=1.10.0=hb80f160_3
libtiff=4.4.0=hfa0b094_4
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.14=h9d8dfc2_4
libxslt=1.1.35=ha8fb468_0
libzlib=1.2.12=h03a7124_3
llvm-openmp=14.0.4=hd125106_0
lxml=4.9.1=py39h9eb174b_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39hb18efdd_1
matplotlib=3.6.0=py39hdf13c20_0
matplotlib-base=3.6.0=py39h35e9e80_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
mplsoccer=1.1.8=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.6.8=pyhd8ed1ab_0
nbconvert=7.0.0=pyhd8ed1ab_0
nbconvert-core=7.0.0=pyhd8ed1ab_0
nbconvert-pandoc=7.0.0=pyhd8ed1ab_0
nbformat=5.5.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
notebook=6.4.12=pyha770c72_0
numpy=1.23.2=py39h3668e8b_0
oauthlib=3.2.1=pypi_0
openjpeg=2.5.0=h5d4e404_1
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1s=h1a28f6b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.4=py39he7125aa_0
pandas-datareader=0.10.0=pyh6c4a22f_0
pandoc=2.12=hca03da5_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.2=py39hca03da5_1
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py39he45c975_2
pip=22.2.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
prompt_toolkit=3.0.31=hd8ed1ab_0
protobuf=3.19.4=py39hfb83b0d_0
psutil=5.9.2=py39h02fc5c5_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyrsistent=0.18.1=py39hb18efdd_1
pysocks=1.7.1=pyha2e5f31_6
python=3.9.13=hc596b02_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.1=pyhd8ed1ab_0
python-slugify=6.1.2=pypi_0
python_abi=3.9=2_cp39
pytz=2022.2.1=pyhd8ed1ab_0
pyyaml=6.0=py39hb18efdd_4
pyzmq=24.0.0=py39h0553236_0
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pyhd8ed1ab_0
scikit-learn=1.1.2=py39h598ef33_0
scipy=1.9.1=py39h737da60_0
seaborn=0.12.0=py39hca03da5_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.3.0=pyhd8ed1ab_1
six=1.16.0=pyh6c4a22f_0
snscrape=0.4.3.20220106=pypi_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.39.3=h2229b38_0
stack_data=0.5.0=pyhd8ed1ab_0
statsmodels=0.13.2=py39h1a28f6b_0
tabulate=0.8.10=py39hca03da5_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-macos=2.10.0=pypi_0
tensorflow-metal=0.6.0=pypi_0
termcolor=2.0.1=pypi_0
terminado=0.15.0=py39h2804cbe_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tornado=6.2=py39h9eb174b_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.4.0=pyhd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022c=h191b570_0
unicodedata2=14.0.0=py39hb18efdd_1
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=4.0.3=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h03a7124_3
zstd=1.5.2=h8128057_4
